feat:output pipe & container service

+++ agent/src/models/container/exec_handle.rs
@@ -6,6 +6,6 @@ use tokio::task::JoinHandle;
 
 pub struct ExecResult {
     pub output: Pin<Box<dyn Stream<Item = Result<LogOutput, bollard::errors::Error>> + Send>>,
-    pub exec_handle: JoinHandle<i64>,
+    pub exec_handle: JoinHandle<i32>,
 }
